Changing project goal from library to simulation server
The goal of harald in no longer to be a ClojureScript library. Although
written in ClojureScript, it's goal is to work as a standalone
application.

Harald will run on node.js and provide a simple REST API, that gives
users of different programming languages access to a running instance of
harald. Harald will also be able to run as a server, that simulates
Bluetooth Low Energy devices.